refactor: [BASKET-39] Layered Architecture 로 구조 변경
