+++ src/main/java/dev/joa/doodledoodlebackend/domain/common/BaseEntity.java
@@ -1,4 +1,4 @@
-package dev.joa.doodledoodlebackend.common.entity;
+package dev.joa.doodledoodlebackend.domain.common;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedBy;
